+++ source/SimulLens_2boxes_V1.2.f90
@@ -76,7 +76,7 @@ program SimulLens
 !--------
 #ifndef halo_only
   ! Arrays on original grid
-  real(8) :: input_map_d(nc,nc)
+  real(4) :: input_map_d(nc,nc)
   real, dimension(nc,nc) :: input_map 
   real, dimension(2*nc,2*nc):: zoom_map
   complex, dimension(nc,nc):: map_cplx
@@ -3032,10 +3032,10 @@ end program SimulLens
     !$omp parallel do default(shared) private(j2,i2,jb,ib,w2,w1,ip,jp)
     do j2=1,n2
        do i2=1,n2
-          jb=shift%y*n1+(j2-1)*frac1+1
-          ib=shift%x*n1+(i2-1)*frac1+1
-          w2=shift%y*n1+(j2-1)*frac1+1-jb
-          w1=shift%x*n1+(i2-1)*frac1+1-ib
+          jb=INT(shift%y*n1+(j2-1)*frac1+1)
+          ib=INT(shift%x*n1+(i2-1)*frac1+1)
+          w2=INT(shift%y*n1+(j2-1)*frac1+1-jb)
+          w1=INT(shift%x*n1+(i2-1)*frac1+1-ib)
           jb=modulo(jb-1,2*n1)+1
           ib=modulo(ib-1,2*n1)+1
           jp=modulo(jb,2*n1)+1
@@ -3131,8 +3131,8 @@ end program SimulLens
     !$omp parallel do default(shared) private(i1,j1,i2,j2,jb,jp,ib,ip,w1,w2,X_array,alpha,coeff)
     do j2=1,n2
        do i2=1,n2
-          jb=shift%y*n1+(j2-1)*frac1+1+defl(i2,j2)%y*n2
-          ib=shift%x*n1+(i2-1)*frac1+1+defl(i2,j2)%x*n2
+          jb=INT(shift%y*n1+(j2-1)*frac1+1+defl(i2,j2)%y*n2)
+          ib=INT(shift%x*n1+(i2-1)*frac1+1+defl(i2,j2)%x*n2)
           jb=modulo(jb-1,2*n1)+1
           ib=modulo(ib-1,2*n1)+1
           jp=modulo(jb,2*n1)+1
